+++ app/helpers/admin/uploads_helper.rb
@@ -10,13 +10,23 @@ module Admin::UploadsHelper
     end
   end
 
-  def get_error_details(error, details)
-    service_name = service.module_parent.to_s.underscore
-    framework = params[:framework].downcase
+  def get_upload_error_row(error)
+    translation_scope = "#{service.module_parent.to_s.underscore}.#{params[:framework].downcase}.admin.uploads.failed"
 
-    return t("#{service_name}.#{framework}.admin.uploads.failed.error_details.#{error}") unless details
+    [
+      {
+        text: t("#{translation_scope}.error_name.#{error[:error]}"),
+      },
+      {
+        text: error[:details] ? get_error_details(translation_scope, error[:error], error[:details]) : t("#{translation_scope}.error_details.#{error[:error]}"),
+      },
+    ]
+  end
 
-    t("#{service_name}.#{framework}.admin.uploads.failed.error_details.#{error}_html", list: details.is_a?(Array) ? details_to_list(details) : details)
+  def get_error_details(translation_scope, error, details)
+    return t("#{translation_scope}.error_details.#{error}") unless details
+
+    t("#{translation_scope}.error_details.#{error}_html", list: details.is_a?(Array) ? details_to_list(details) : details)
   end
 
   def details_to_list(details)