+++ app/services/facilities_management/rm6378/admin/files_importer.rb
@@ -1,0 +1,2 @@
+class FacilitiesManagement::RM6378::Admin::FilesImporter < FilesImporter
+end